bronze_level/requirement.txt
appdirs==1.4.4
asgiref==3.8.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.7.4
charset-normalizer==3.3.2
cssselect==1.2.0
Django==5.0.7
fake-useragent==1.5.1
feedparser==6.0.11
frozendict==2.4.4
greenlet==3.0.3
html5lib==1.1
idna==3.7
importlib_metadata==8.2.0
lxml==5.2.2
multitasking==0.0.11
numpy==2.0.1
pandas==2.2.2
parse==1.20.2
peewee==3.17.6
platformdirs==4.2.2
prometheus_client==0.20.0
psycopg2-binary==2.9.9
pyee==11.1.0
pyppeteer==2.0.0
pyquery==2.0.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
requests-html==0.10.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.31
sqlparse==0.5.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.19
w3lib==2.2.1
webencodings==0.5.1
websockets==10.4
yahoo-fin==0.8.9.1
yfinance==0.2.41
zipp==3.19.2
